Issue28 81

+++ Museum-Backend/src/main/java/ca/mcgill/ecse321/museum/service/TimePeriodService.java
@@ -1,9 +1,6 @@
 package ca.mcgill.ecse321.museum.service;
 
-import ca.mcgill.ecse321.museum.dao.ScheduleOfTimePeriodRepository;
 import ca.mcgill.ecse321.museum.dao.TimePeriodRepository;
-import ca.mcgill.ecse321.museum.model.Schedule;
-import ca.mcgill.ecse321.museum.model.ScheduleOfTimePeriod;
 import ca.mcgill.ecse321.museum.model.TimePeriod;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,8 +20,6 @@ import java.util.List;
 public class TimePeriodService {
   @Autowired
   private TimePeriodRepository timePeriodRepository;
-  @Autowired
-  private ScheduleOfTimePeriodRepository scheduleOfTimePeriodRepository;
 
   // GET
 
@@ -32,13 +27,16 @@ public class TimePeriodService {
    * Method to get a time period from database
    *
    * @param timePeriodId - id of time period
-   * @return
+   * @return time period
    * @author VZ
    */
-
   @Transactional
   public TimePeriod getTimePeriod(long timePeriodId) {
-    return timePeriodRepository.findTimePeriodByTimePeriodId(timePeriodId);
+    TimePeriod timePeriod = timePeriodRepository.findTimePeriodByTimePeriodId(timePeriodId);
+    if (timePeriod == null) {
+      throw new IllegalArgumentException("Time period does not exist");
+    }
+    return timePeriod;
   }
 
   // CREATE
@@ -46,12 +44,11 @@ public class TimePeriodService {
   /**
    * Create a TimePeriod and save to database
    *
-   * @param startDate
-   * @param endDate
-   * @return
+   * @param startDate - start date of time period
+   * @param endDate   - end date of time period
+   * @return time period
    * @author VZ
    */
-
   @Transactional
   public TimePeriod createTimePeriod(Timestamp startDate, Timestamp endDate) {
 
@@ -76,13 +73,15 @@ public class TimePeriodService {
   /**
    * Delete a TimePeriod from database by ID
    *
-   * @param timePeriodId
+   * @param timePeriodId - id of time period
    * @author VZ
    */
-
   @Transactional
   public void deleteTimePeriod(long timePeriodId) {
-
+    TimePeriod timePeriod = timePeriodRepository.findTimePeriodByTimePeriodId(timePeriodId);
+    if (timePeriod == null) {
+      throw new IllegalArgumentException("Time period does not exist");
+    }
     timePeriodRepository.deleteById(timePeriodId);
   }
 
@@ -91,52 +90,58 @@ public class TimePeriodService {
   /**
    * Edit a TimePeriod by ID and save to database
    *
-   * @param timePeriodId
-   * @param startDate
-   * @param endDate
-   * @return
+   * @param timePeriodId - id of time period
+   * @param startDate    - start date of time period
+   * @param endDate      - end date of time period
+   * @return edited time period
    * @author VZ
    */
-
   @Transactional
   public TimePeriod editTimePeriod(long timePeriodId, Timestamp startDate, Timestamp endDate) {
-    // input validation
-    if (startDate == null || endDate == null) {
-      throw new IllegalArgumentException("Start date and end date cannot be null");
-    }
+
     if (startDate.after(endDate)) {
       throw new IllegalArgumentException("Start date cannot be after end date");
     }
     // find TimePeriod
     TimePeriod timePeriod = timePeriodRepository.findTimePeriodByTimePeriodId(timePeriodId);
     if (timePeriod == null) {
-      throw new IllegalArgumentException("TimePeriod does not exist");
+      throw new IllegalArgumentException("Time period does not exist");
+    }
+
+    if (startDate == null || endDate == null) {
+      throw new IllegalArgumentException("Dates cannot be null");
     }
     // edit TimePeriod
     timePeriod.setStartDate(startDate);
     timePeriod.setEndDate(endDate);
+
     timePeriodRepository.save(timePeriod);
     return timePeriod;
   }
 
   /**
-   * Helper method to get the time periods of a schedule
+   * Get all time periods from database
    *
-   * @param schedule
-   * @return
+   * @return list of time periods
    * @author VZ
    */
+  @Transactional
+  public List<TimePeriod> getAllTimePeriods() {
+    return toList(timePeriodRepository.findAll());
+  }
 
-  public List<TimePeriod> getTimePeriodsOfSchedule(Schedule schedule) {
-
-    List<ScheduleOfTimePeriod> scheduleOfTimePeriods = scheduleOfTimePeriodRepository
-        .findScheduleOfTimePeriodBySchedule(schedule);
-
-    List<TimePeriod> timePeriods = new ArrayList<TimePeriod>();
-    for (ScheduleOfTimePeriod scheduleOfTimePeriod : scheduleOfTimePeriods) {
-      timePeriods.add(scheduleOfTimePeriod.getTimePeriod());
+  /**
+   * Method to convert an Iterable to a List
+   *
+   * @param iterable - Iterable
+   * @return List
+   * @author From tutorial notes
+   */
+  private <T> List<T> toList(Iterable<T> iterable) {
+    List<T> resultList = new ArrayList<T>();
+    for (T t : iterable) {
+      resultList.add(t);
     }
-
-    return timePeriods;
+    return resultList;
   }
 }
